feat: Correção na tela de dashboard
